requirements.txt
discord.py
python-dotenv
beautifulsoup4
requests
cloudscraper
Pillow
aiosqlite
